second update...
1. better error checking... i was trying a whole bunch of different
media formats, interaction scenarios, etc., and think i finally have
all the bulletproofing in place@

2. on my system, albumArtURI always comes back as a URI not a path -
also i moved the DIDL parsing to a seperate function so it can be used
in other places (like my code!)

3. add track position and duration. duration is accurate. position
appears to be updated every 5s.

4. added functions to get/set the name of the speaker, the play mode,
skip to a location, and flush the queue.